Add Check Ensuring CUs Cover the Action
Previously, it was possible to submit a transaction without compliance
units, trivial delta proof, and it would pass given that the resource
logics pass. That means, you could e.g. consume resources without ever
proving they existed before.

This commit changes that by checking that the number of tags in an
action in exactly twice the number of compliance units.

Coupled with the check ensuring no double-creation/consumption in the
same action it grants the desired property.